refactor: extract app shell, terminal lifecycle, and process helpers
